fix: Show settings on iOS This change includes a drive-by refactor to move many iOS views into their own files.
